Fixed obscure pid file bug. If atoi() returns <= 0, kill() doesn't treat that in the expected way.

+++ code/sys/sys_unix.c
@@ -773,5 +773,11 @@ Sys_PIDIsRunning
 */
 qboolean Sys_PIDIsRunning( int pid )
 {
-	return kill( pid, 0 ) == 0;
+	// For kill(), pid <= 0 has special meanings so
+	// we can't really use it to check if something
+	// is actually running.
+	if( pid <= 0 )
+		return qfalse;
+	else
+		return kill( pid, 0 ) == 0;
 }
